adding Adapter types for near detector
